style: tambah components table
Deskripsi:
- Menambahkan Components Table
- update table step
- update table swot
- update table mlai

No Issue
