+++ src/modules/mlai/back/component/table_mlai.tsx
@@ -24,9 +24,10 @@ import {
 } from "react-icons/md";
 import { isModalMlAi } from "../val/val_mlai";
 import ModalDelMlAi from "./modal_del_mlai";
-import { CiRead } from "react-icons/ci";
+import { CiRead, CiUnread } from "react-icons/ci";
 import { useDisclosure } from "@mantine/hooks";
 import { useState } from "react";
+import ComponentTable from "./component_table";
 
 /**
  * Fungsi untuk menampilkan table ml ai.
@@ -116,54 +117,7 @@ export default function TableMlAi() {
                     </Table.Tr>
                   </Table.Thead>
                   {elements.map((v, i) => (
-                    <Table.Tbody key={i}>
-                      <Table.Tr>
-                        <Table.Td>{v.no}</Table.Td>
-                        <Table.Td>{v.name}</Table.Td>
-                        <Table.Td>
-                          <Center>
-                            <ActionIcon
-                              variant="transparent"
-                              color="rgba(5, 128, 23, 1)"
-                              size="xl"
-                              aria-label="Edit"
-                              onClick={toggle}
-                            >
-                              <CiRead size={20} />
-                            </ActionIcon>
-                            <ActionIcon
-                              variant="transparent"
-                              color="rgba(5, 128, 23, 1)"
-                              size="xl"
-                              aria-label="Edit"
-                              onClick={() =>
-                                router.push("ml-ai/edit/IKomangAyu")
-                              }
-                            >
-                              <MdEditCalendar size={20} />
-                            </ActionIcon>
-                            <ActionIcon
-                              variant="transparent"
-                              color="rgba(209, 4, 4, 1)"
-                              size="xl"
-                              aria-label="Delete"
-                              onClick={() => setOpenModal(true)}
-                            >
-                              <MdDelete size={20} />
-                            </ActionIcon>
-                          </Center>
-                        </Table.Td>
-                      </Table.Tr>
-                      <Table.Tr
-                        style={{
-                          borderBottom: "1px solid #CED4D9",
-                        }}
-                      >
-                        <Table.Td colSpan={5}>
-                          <ClickMore v={v} />
-                        </Table.Td>
-                      </Table.Tr>
-                    </Table.Tbody>
+                    <ComponentTable v={v} i={i} key={i} />
                   ))}
                 </Table>
               </ScrollArea>
@@ -188,36 +142,4 @@ export default function TableMlAi() {
   );
 }
 
-function ClickMore({ v }: { v: any }) {
-  const open = useState(false);
 
-  return (
-    <>
-      <Stack>
-      <Group>
-          <ActionIcon onClick={() => open[1](!open[0])}>
-            {open[0] ? <MdArrowCircleDown /> : <MdArrowCircleUp />}
-          </ActionIcon>
-        </Group>
-        <Collapse
-          in={open[0]}
-          transitionDuration={500}
-          transitionTimingFunction="linear"
-        >
-          <Box
-            style={{
-              backgroundColor: "gray",
-              padding: 20,
-              borderRadius: 10,
-            }}
-          >
-            <Text c={"white"} fw={"bold"} fz={20} mb={10}>
-              Content
-            </Text>
-            <Text c={"white"}>{v.content}</Text>
-          </Box>
-        </Collapse>
-      </Stack>
-    </>
-  );
-}
