+++ src/modules/step/back/val/val_step.ts
@@ -3,3 +3,4 @@
 import { atom } from "jotai"
 
 export const isModalStep = atom(false)
+export const isModalvalueStep = atom(false)
