+++ src/modules/swot/back/component/table_swot.tsx
@@ -22,6 +22,7 @@ import ModalDelSwot from "./modal_del_swot";
 import { CiRead, CiUnread } from "react-icons/ci";
 import { useDisclosure } from "@mantine/hooks";
 import { useState } from "react";
+import ComponentTable from "./component_table";
 
 /**
  * Fungsi untuk menampilkan view table swot.
@@ -119,75 +120,7 @@ export default function TableSwot() {
                     </Table.Tr>
                   </Table.Thead>
                   {elements.map((v, i) => (
-                    <Table.Tbody key={i}>
-                      <Table.Tr>
-                        <Table.Td>{v.no}</Table.Td>
-                        <Table.Td>{v.name}</Table.Td>
-                        <Table.Td>{v.category}</Table.Td>
-                        {/* <Table.Td>{v.content}</Table.Td> */}
-                        <Table.Td>
-                          <Center>
-                            <ActionIcon
-                              variant="transparent"
-                              color="rgba(5, 128, 23, 1)"
-                              size="xl"
-                              aria-label="Edit"
-                              onClick={() => open[1](!open[0])}
-                            >
-                              {open[0] ? <CiRead size={20} /> : <CiUnread/>}
-                               
-                            </ActionIcon>
-                            <ActionIcon
-                              variant="transparent"
-                              color="rgba(5, 128, 23, 1)"
-                              size="xl"
-                              aria-label="Edit"
-                              onClick={() =>
-                                router.push("swot/edit/IKomangAyu")
-                              }
-                            >
-                              <MdEditCalendar size={20} />
-                            </ActionIcon>
-                            <ActionIcon
-                              variant="transparent"
-                              color="rgba(209, 4, 4, 1)"
-                              size="xl"
-                              aria-label="Delete"
-                              onClick={() => setOpenModal(true)}
-                            >
-                              <MdDelete size={20} />
-                            </ActionIcon>
-                          </Center>
-                        </Table.Td>
-                      </Table.Tr>
-                      <Table.Tr
-                        style={{
-                          borderBottom: "1px solid #CED4D9",
-                        }}
-                      >
-                        <Table.Td colSpan={5}>
-                          {/* <Collapse
-                            in={open[0]}
-                            transitionDuration={500}
-                            transitionTimingFunction="linear"
-                          >
-                            <Box
-                              style={{
-                                backgroundColor: "gray",
-                                padding: 20,
-                                borderRadius: 10,
-                              }}
-                            >
-                              <Text c={"white"} fw={"bold"} fz={20} mb={10}>
-                                Content
-                              </Text>
-                              <Text c={"white"}>{v.content}</Text>
-                            </Box>
-                          </Collapse> */}
-                          <ClickMore v={v} />
-                        </Table.Td>
-                      </Table.Tr>
-                    </Table.Tbody>
+                    <ComponentTable v={v} i={i} key={i} />
                   ))}
                 </Table>
               </ScrollArea>
